requirements.txt
auto_mix_prep==0.2.0
matplotlib==3.5.0
numpy==1.20.3
pandas==1.3.5
Pillow==9.1.0
pycocotools==2.0
PyYAML==6.0
scikit_image==0.18.3
skimage==0.0
swin_transformer_v2==0.1
timm==0.5.4
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
